refactor(base): Create the `e2ee::decrypt::sync_timeline_event` response processor.
This patch extracts the `BaseClient::decrypt_sync_room_event` method
into the new `e2ee::decrypt::sync_timeline_event` response processor.
